requirements.txt
Flask==2.3.2
requests==2.31.0
Pillow==12.0.0
numpy==1.26.2
faiss-cpu==1.12.0
transformers==4.35.0
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87


